Checkpoint: added CMD_GET_FUOTA_METADATA and logic to fetch metadata from flash
Checkpoint: Added a .conf file to enable the CRYPTO and MBEDTLS functionality into the nrf5340dk board build of the `hw_modem` application. The code builds and runs, but the hash functions have not yet been added.

Checkpoint: writing FUOTA fragments to flash and loading them on CMD_GET_FUOTA_METADATA

Checkpoint: calculation of SHA256 hash of received FUOTA fragments on the device